+++ test/PHPUnit/connecteur/transformation-generique/TransformationGeneriqueTest.php
@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 class TransformationGeneriqueTest extends PastellTestCase
 {
     /**
@@ -17,8 +19,8 @@ class TransformationGeneriqueTest extends PastellTestCase
 
         $connecteurConfig->addFileFromCopy(
             'definition',
-            "definition.json",
-            __DIR__ . "/fixtures/definition.json"
+            'definition.json',
+            __DIR__ . '/fixtures/definition.json'
         );
         /** @var TransformationGenerique $connector */
         $connector = $this->getConnecteurFactory()->getConnecteurById($id_ce);
@@ -27,27 +29,64 @@ class TransformationGeneriqueTest extends PastellTestCase
 
     /**
      * @throws DonneesFormulaireException
+     * @throws JsonException
+     * @throws UnrecoverableException
      */
-    public function testExtraction()
+    public function testExtraction(): void
     {
         $transformationGenerique = $this->getConnecteur();
         $donneesFormulaire = $this->getDonneesFormulaireFactory()->getNonPersistingDonneesFormulaire();
         $transformationGenerique->transform($donneesFormulaire);
-        $this->assertEquals("bar", $donneesFormulaire->get('foo'));
+        static::assertSame('bar', $donneesFormulaire->get('foo'));
     }
 
     /**
      * @throws DonneesFormulaireException
+     * @throws JsonException
+     * @throws UnrecoverableException
      */
-    public function testTestTransform()
+    public function testTestTransform(): void
     {
         $transformationGenerique = $this->getConnecteur();
         $donneesFormulaire = $this->getDonneesFormulaireFactory()->getNonPersistingDonneesFormulaire();
         $info = $this->createDocument('test');
         $donneesFormulaire->id_d = $info['id_d'];
-        $this->assertEquals(
+        static::assertSame(
             '{"foo":"bar","envoi_signature":"true","titre":"Ceci est mon titre","from_pa_metadata":"Bourg-en-Bresse Eric"}',
             $transformationGenerique->testTransform($donneesFormulaire)
         );
+    }
+
+    /**
+     * @throws DonneesFormulaireException
+     * @throws JsonException
+     * @throws UnrecoverableException
+     * @throws Exception
+     */
+    public function testTransformWithData(): void
+    {
+        $connectorId = $this->createConnector(
+            'transformation-generique',
+            'Transformation generique'
+        )['id_ce'];
+
+        $connectorConfig = $this->getDonneesFormulaireFactory()->getConnecteurEntiteFormulaire($connectorId);
+
+        $connectorConfig->addFileFromCopy(
+            'definition',
+            'definition.json',
+            __DIR__ . '/fixtures/definition_with_file.json',
+        );
+            $connectorConfig->addFileFromCopy(
+                'data',
+                'data.json',
+                __DIR__ . '/fixtures/data.json',
+            );
+        /** @var TransformationGenerique $connector */
+        $connector = $this->getConnecteurFactory()->getConnecteurById($connectorId);
+
+        $donneesFormulaire = $this->getDonneesFormulaireFactory()->getNonPersistingDonneesFormulaire();
+        $connector->transform($donneesFormulaire);
+        static::assertSame('valueFromFile', $donneesFormulaire->get('foo'));
     }
 }